slice of life updates; renaming files
