+++ src/components/Home-Page.js
@@ -6,11 +6,11 @@ import Pic1 from "./img/landingPic1.png";
 import Baby from "./img/landingBaby.png";
 import Parents from "./img/landingParents.png";
 import { useNavigate } from "react-router-dom";
-import { getMap } from "./API";
+import { getMap } from "./GooglePlacesAPI";
 
 const Home = () => {
   const navigate = useNavigate();
-  getMap();
+  getMap(); //calling getMap function for API
 
   return (
     <div
